Cluster-Analysis
Cluster Analysis on Health Outcome of Different Countries using R Statistical Software Application.

In this project I discovered subgroups of observations within variables. I cluster the factors as outcomes in the same group to be similar and observations in different groups to be dissimilar.

This being an unsupervised learning method, which implies that it seeks to find relationships between the n observations without being trained by a response variable.

K-Mean and Hierarchical methods was used to clearly view and understand the similar variables.